CRQ-TOE-009: Leech lattice + V♮ moonshine module hash tests
- 5 tests: Leech constants, V♮ structure, j-coefficients, McKay-Thompson, Griess=VOA
- All hashed through Borcherds cell (8,4,3) and Monster-complete (71,59,47)
- Fixed u16 overflow in bott() for Monster-complete cell
- 196883 and 196884 hash to distinct orbifold points ✓
- 15 conjugacy classes all distinct ✓

+++ src/da51_macros.rs
@@ -39,7 +39,7 @@ impl GriessCell {
     }
 
     pub const fn bott(&self) -> u8 {
-        ((self.n as u16 * self.m as u16 * self.c as u16) % 8) as u8
+        ((self.n as u32 * self.m as u32 * self.c as u32) % 8) as u8
     }
 
     pub const fn hecke_index(&self) -> usize {
